requirements.txt
Django==3.2.10
djangorestframework==3.13.1
drf-yasg==1.20.0
django-storages[google]==1.12.3
django-environ==0.8.1
django-cors-headers==3.10.1
psycopg2-binary==2.9.2
gunicorn==20.1.0
google-cloud-secret-manager==2.8.0
